Added non structured response support in do_query
do_query always returned the structured response..
I have modified the do_query method, so that if :fmt is not provided in parameters, then non-structured response will be returned. Column names will be used as keys.

Please note that quickbase API converts special characters and space in column names with '_', and this method will return it as is.